requirements.txt
numba>=0.52.0
librosa>=0.9.1
numpy>=1.19.2
tqdm>=4.59.0
visualdl>=2.1.1
resampy==0.2.2
soundfile>=0.12.1
soundcard>=0.4.2
pyyaml>=5.4.1
termcolor>=1.1.0
typeguard==2.13.3
scikit-learn>=1.0.2
pydub>=0.25.1
torchaudio>=0.12.1
torchinfo>=1.7.2
av>=10.0.0